Fix bug with tritmap integration

+++ kernel/src/math.h
@@ -22,6 +22,6 @@ uint64_t power_uint8(const uint8_t base, const uint8_t exponent) {
 
 // Ceil function
 uint64_t ceil(const uint32_t x, const uint32_t y) {
-    uint64_t z = x;
+    uint64_t z = (uint64_t)x;
     return (z + y - 1) / y;
 }
